requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
blinker==1.6.3
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
cryptography==41.0.4
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
firebase-admin==6.2.0
Flask==3.0.0
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud==0.34.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.4.0
jupyter_core==5.4.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
msgpack==1.0.7
nest-asyncio==1.5.8
packaging==23.2
paho-mqtt==1.6.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
requests==2.31.0
rsa==4.9
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.11.2
uritemplate==4.1.1
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
